requirements.txt
streamlit>=1.32
rdkit-pypi
scikit-learn>=1.4
lightgbm>=4.0
xgboost>=2.0
pandas>=2.0
numpy>=1.26
plotly>=5.18
openpyxl>=3.1
imbalanced-learn>=0.12
shap>=0.44
joblib>=1.3
scipy>=1.12
